Update BraveShields to depend on Data, remove BraveShields dep on Data.
