+++ ios/brave-ios/Sources/BrowserMenu/Identifiers.swift
@@ -3,6 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
+import BraveShields
 import BraveWallet
 import Foundation
